Json extraction - look for the last ';' instead of the first one
My address has a ';' in it in the eyeonwater system, and this string slicing fails. Looking for the last ';' would work better, so we get the end of the javascript line instead of a random string slice ending in the middle of the address, that later fails the json decoding.